+++ mex/editor/components.py
@@ -1,17 +1,8 @@
-from collections.abc import Callable
-from typing import Any, cast
+from typing import Any
 
 import reflex as rx
-from reflex.vars import Var
 
-from mex.editor.ingest.state import IngestState
-from mex.editor.models import (
-    PaginationButtonOptions,
-    PaginationOptions,
-    PaginationPageOptions,
-)
 from mex.editor.rules.models import EditorValue
-from mex.editor.search.state import SearchState
 from mex.editor.state import State
 
 
@@ -51,7 +42,7 @@ def render_additional_titles(titles: list[EditorValue]) -> rx.Component:
             rx.hover_card.trigger(
                 rx.badge(
                     State.label_additional_titles,
-                    style=rx.Style(margin="auto 0"),
+                    style=rx.Style(margin="auto 0", cursor="default"),
                     custom_attrs={"data-testid": "additional-titles-badge"},
                 ),
                 custom_attrs={"data-testid": "tooltip-additional-titles-trigger"},
@@ -153,84 +144,6 @@ def render_badge(text: str | None) -> rx.Component:
         variant="soft",
         color_scheme="gray",
         style=rx.Style(margin="auto 0"),
-    )
-
-
-def pagination_abstract(options: PaginationOptions, **kwargs: dict) -> rx.Component:
-    """Create pagination based on given options."""
-    style = rx.Style().update(kwargs.pop("style", None))
-    return rx.flex(
-        rx.button(
-            rx.text(State.label_pagination_previous_button),
-            on_click=options.prev_options.on_click,
-            disabled=options.prev_options.disabled,
-            variant="surface",
-            custom_attrs={"data-testid": "pagination-previous-button"},
-            style=rx.Style(minWidth="10%"),
-        ),
-        rx.select(
-            options.page_options.pages,
-            value=options.page_options.current_page.to_string()
-            if isinstance(options.page_options.current_page, Var)
-            else f"{options.page_options.current_page}",
-            on_change=options.page_options.on_change,
-            disabled=options.page_options.disabled,
-            custom_attrs={"data-testid": "pagination-page-select"},
-        ),
-        rx.button(
-            rx.text(State.label_pagination_next_button, weight="bold"),
-            on_click=options.next_options.on_click,
-            disabled=options.next_options.disabled,
-            variant="surface",
-            custom_attrs={"data-testid": "pagination-next-button"},
-            style=rx.Style(minWidth="10%"),
-        ),
-        spacing="4",
-        style=style,
-    )
-
-
-def pagination(
-    state: type[IngestState | SearchState], *page_load_hooks: Callable[[], Any]
-) -> rx.Component:
-    """Create pagination for IngestState or SearchState."""
-    current_page = cast("rx.Var[int]", state.current_page)
-
-    return pagination_abstract(
-        PaginationOptions(
-            PaginationButtonOptions(
-                state.disable_previous_page,
-                [
-                    state.go_to_previous_page,
-                    state.scroll_to_top,
-                    state.refresh,
-                    state.resolve_identifiers,
-                    *page_load_hooks,
-                ],
-            ),
-            PaginationButtonOptions(
-                state.disable_next_page,
-                [
-                    state.go_to_next_page,
-                    state.scroll_to_top,
-                    state.refresh,
-                    state.resolve_identifiers,
-                    *page_load_hooks,
-                ],
-            ),
-            PaginationPageOptions(
-                current_page,
-                state.page_selection,
-                state.disable_page_selection,
-                [
-                    state.set_current_page,
-                    state.scroll_to_top,
-                    state.refresh,
-                    state.resolve_identifiers,
-                    *page_load_hooks,
-                ],
-            ),
-        )
     )
