+++ mex/editor/search/state.py
@@ -16,7 +16,7 @@ from mex.editor.exceptions import escalate_error
 from mex.editor.label_var import label_var
 from mex.editor.locale_service import LocaleService
 from mex.editor.models import SearchResult
-from mex.editor.pagination_state_mixin import PaginationStateMixin
+from mex.editor.pagination_component import PaginationStateMixin
 from mex.editor.search.models import (
     ReferenceFieldFilter,
     ReferenceFieldIdentifierFilter,
